spliting out PixmapItem class into its own file
